requirements.txt
openai>=1.12.0
pydub>=0.25.1
click>=8.1.0
rich>=13.7.0
python-dotenv>=1.0.0
tenacity>=8.2.3
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
